# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Innovo Labs Digital Initiatives Portal API
  version: "1.0.0"
  description: >
    REST API for the Innovo Labs portal. All endpoints are Next.js Route Handlers
    under `/api/**`, running as serverless functions (Node.js runtime) on Vercel.

    Authentication: session is a JWT (HS256, jose) stored in an HttpOnly cookie
    named `innovo_session`, issued after Microsoft Entra ID (OAuth2/OIDC) SSO.
    Because the cookie is HttpOnly it is sent automatically by the browser; there
    is no bearer token in the Authorization header.

    Authorization: enforced in two layers — the `defineRoute` wrapper (role gate +
    Zod validation + rate limit) and PostgreSQL Row-Level Security.

    Common responses (returned by the shared wrapper):
      401 Not authenticated · 403 Forbidden · 400 validation error ·
      429 Too Many Requests · 500 Internal server error.
servers:
  - url: https://<your-vercel-domain>
    description: Production (Vercel)
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: innovo_session
      description: HttpOnly JWT session cookie issued after Entra ID SSO.
  responses:
    Unauthorized:
      description: No valid session cookie.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
security:
  - sessionCookie: []
paths:
...
  /api/auth/callback/azure:
    get:
      tags: [Auth]
      security: []
      summary: SSO callback
      description: Validates `state` against the cookie, exchanges the code, issues the session cookie.
      parameters:
        - { in: query, name: code, schema: { type: string }, required: true }
        - { in: query, name: state, schema: { type: string }, required: true }
      responses:
        "302": { description: Redirect to the app (or /login with an error param). }
  /api/auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Programmatic login (non-SSO paths)
      responses:
        "200": { description: Session established. }
        "401": { $ref: '#/components/responses/Unauthorized' }
...